% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ranges.R
\name{get_annotation}
\alias{get_annotation}
\title{Extract gene annotation as a \code{GRanges} object}
\usage{
get_annotation(
  species,
  transcripts = c("all", "longest"),
  features = c("all", "exons")
)
}
\arguments{
\item{species}{Character vector of PLAZA species IDs.}

\item{transcripts}{Character indicating what genes for which ranges will
be extracted. One of 'all' (for each gene, all transcripts and
associated features), or 'longest' (for each gene, only longest transcripts
and associated features).}

\item{features}{Character indicating what features to extract. One of
'all' (all features, including exons, mRNAs, CDS, UTRs, etc), or
'exons' (only exon features).}
}
\value{
A \code{GRanges} object for a single species, or a \code{GRangesList} object
for multiple species.
}
\description{
Extract gene annotation as a \code{GRanges} object
}
\examples{
ranges <- get_annotation(
    species = c("aly", "ath"), 
    transcripts = "longest", features = "all"
)
}
