% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_homology.R
\name{get_block_duplicates}
\alias{get_block_duplicates}
\title{Get anchor pairs in intraspecies syntenic regions}
\usage{
get_block_duplicates(species)
}
\arguments{
\item{species}{Character vector of PLAZA species IDs.}
}
\value{
A data frame (or list of data frames, in case of multiple species)
with the following variables:
\describe{
\item{multiplicon_id}{Numeric, multiplicon ID.}
\item{anchor1}{Character, gene ID of anchor 1.}
\item{anchor2}{Character, gene ID of anchor 2.}
\item{Ks}{Numeric, synonymous substitution rates. If not calculated,
NA is used.}
\item{level}{Numeric, multiplicon level.}
}
}
\description{
Get anchor pairs in intraspecies syntenic regions
}
\examples{
sd <- get_block_duplicates("lpe")
}
