% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_descriptions}
\alias{get_descriptions}
\title{Get short, one-line gene descriptions}
\usage{
get_descriptions(species)
}
\arguments{
\item{species}{Character vector of PLAZA species IDs.}
}
\value{
A 2-column data frame with columns \strong{gene} (character)
and \strong{description} (character).
}
\description{
Get short, one-line gene descriptions
}
\examples{
des <- get_descriptions("ota")
}
