% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_homology.R
\name{get_family_assignments}
\alias{get_family_assignments}
\title{Get gene family assignments for all genes in a PLAZA instance}
\usage{
get_family_assignments(
  instance = c("Dicots", "Monocots", "Diatoms", "Pico"),
  type = c("homo", "ortho")
)
}
\arguments{
\item{instance}{Character scalar with name of PLAZA instance for which
gene family assignments will be extracted.
One of 'Dicots', 'Monocots', 'Diatoms', or 'Pico'.}

\item{type}{Character indicating the type of gene family to use. One of
'homo' (homologous gene families) or 'ortho' (orthologous gene families).}
}
\value{
A data frame (or list of data frames, in case of multiple species)
with the following variables:
\describe{
\item{family_id}{Character, gene family ID.}
\item{species}{Character, species ID.}
\item{gene_id}{Character, gene ID.}
}
}
\description{
Get gene family assignments for all genes in a PLAZA instance
}
\examples{
fams <- get_family_assignments("Diatoms", type = "ortho")
}
