% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_homology.R
\name{get_family_functions}
\alias{get_family_functions}
\title{Get enriched functional terms in gene families of a PLAZA instance}
\usage{
get_family_functions(
  instance = c("Dicots", "Monocots"),
  type = c("homo", "ortho")
)
}
\arguments{
\item{instance}{Character scalar with name of PLAZA instance for which
gene family functions will be extracted.
One of 'Dicots' or 'Monocots' (not available for 'Diatoms' and 'Pico').}

\item{type}{Character indicating the type of gene family to use. One of
'homo' (homologous gene families) or 'ortho' (orthologous gene families).}
}
\value{
A data frame (or list of data frames, in case of multiple species)
with the following variables:
\describe{
\item{family_id}{Character, gene family ID.}
\item{term_id}{Character, functional term ID.}
\item{term_description}{Character, functional term description.}
\item{ontology}{Character, ontology from where term comes.}
\item{score}{Numeric, enrichment score.}
}
}
\description{
Get enriched functional terms in gene families of a PLAZA instance
}
\examples{
fam_func <- get_family_functions("Dicots", type = "ortho")
}
