% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functional_annotation.R
\name{get_functional_annotation}
\alias{get_functional_annotation}
\title{Get mapping between alternative IDs}
\usage{
get_functional_annotation(species, ontology = c("GO", "MapMan", "InterPro"))
}
\arguments{
\item{species}{Character vector of PLAZA species IDs.}

\item{ontology}{Character indicating what ontology to use to extract
functional annotation. One of 'GO' (Gene Ontology), 'MapMan',
or 'InterPro'.}
}
\value{
A data frame with functional annotation terms for all genes.
}
\description{
Get mapping between alternative IDs
}
\examples{
fannot <- get_functional_annotation("ota", "MapMan")
}
