% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{get_genome}
\alias{get_genome}
\title{Load genome sequences as a \code{DNAStringSet} object}
\usage{
get_genome(species)
}
\arguments{
\item{species}{Character vector of PLAZA species IDs.}
}
\value{
A \code{DNAStringSet} object for a single species, or a list
of \code{DNAStringSet} objects for multiple species.
}
\description{
Load genome sequences as a \code{DNAStringSet} object
}
\examples{
species <- "cre"
seq <- get_genome(species)
}
