% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_id_conversions}
\alias{get_id_conversions}
\title{Get mapping between alternative IDs}
\usage{
get_id_conversions(species)
}
\arguments{
\item{species}{Character vector of PLAZA species IDs.}
}
\value{
A 2-column data frame with columns \strong{original_id} (character),
\strong{alt_id_type} (character), and \strong{alt_id} (character).
}
\description{
Get mapping between alternative IDs
}
\examples{
ids <- get_id_conversions("ota")
}
