% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sequences.R
\name{get_sequences}
\alias{get_sequences}
\title{Get sequences for genomic loci as a \code{DNAStringSet} object}
\usage{
get_sequences(
  species,
  type = c("CDS", "transcript", "protein"),
  transcripts = c("all", "longest")
)
}
\arguments{
\item{species}{Character vector of PLAZA species IDs.}

\item{type}{Character indicating what type of locus to extract.
One of 'CDS', 'transcript', or 'protein'.}

\item{transcripts}{Character indicating what genes for which sequences will
be extracted. One of 'all' (for each gene, all transcripts and
associated features), or 'longest' (for each gene, only longest transcripts
and associated features).}
}
\value{
A \code{DNAStringSet} object for a single species, or a list
of \code{DNAStringSet} objects for multiple species.
}
\description{
Get sequences for genomic loci as a \code{DNAStringSet} object
}
\examples{
species <- "cre"
seq <- get_sequences(species, type = "protein", transcripts = "longest") 
}
