% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_tx2gene}
\alias{get_tx2gene}
\title{Get transcript-to-gene mapping}
\usage{
get_tx2gene(species, transcripts = c("all", "longest"))
}
\arguments{
\item{species}{Character vector of PLAZA species IDs.}

\item{transcripts}{Character indicating what transcripts to use for each
gene. One of 'all' (all transcripts for each gene), or 'longest'
(only longest transcript for each gene).}
}
\value{
A 2-column data frame with columns
\strong{transcript_id} (character) and \strong{gene_id} (character)
}
\description{
Get transcript-to-gene mapping
}
\examples{
species <- "ota"
tx2gene <- get_tx2gene(species, "all")
}
