% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plaza_tree}
\alias{plaza_tree}
\title{Phylogenetic tree with relationships between species in all PLAZA instances}
\format{
A \code{phylo} object with tip labels corresponding to PLAZA species IDs.
Lengths of all branches are 0.05 and do not represent any meaningful
measure of time (neither absolute not accumulated substitutions).
}
\usage{
data(plaza_tree)
}
\description{
Trees were obtained from the home page of each PLAZA instance and combined
to include all (non-redundant) species.
}
\examples{
data(plaza_tree)
}
\keyword{datasets}
